Refactored views
Fixed add-courses page's search not including subjects (again)
Disabled DRF
