+++ django_project/roster_sniper/settings/base.py
@@ -14,7 +14,6 @@ INSTALLED_APPS = [
 	'users.apps.UsersConfig',
 
 	'crispy_forms',
-	'rest_framework',
 
 	'django.contrib.admin',
 	'django.contrib.auth',
